Updated Example App - moved important bits into context - factored out components for maintainability
